Se corrigio la logica de eliminar torneo para que no se permita si existen equipos asociados

+++ src/main/java/com/tallerwebi/dominio/ServicioTorneoImpl.java
@@ -2,6 +2,7 @@ package com.tallerwebi.dominio;
 
 import com.tallerwebi.dominio.excepcion.FechaIncoherenteException;
 import com.tallerwebi.dominio.excepcion.FechasSuperpuestasException;
+import com.tallerwebi.dominio.excepcion.NoSePuedeEliminarUnTorneoSiTieneEquiposAsociadosException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,12 +15,14 @@ import java.util.List;
 public class ServicioTorneoImpl implements ServicioTorneo {
 
     private RepositorioTorneo repositorioTorneo;
+    private RepositorioEquipo repositorioEquipo;
 
     @Autowired
     public ServicioTorneoImpl(
-            RepositorioTorneo repositorioTorneo) {
-
+            RepositorioTorneo repositorioTorneo,
+            RepositorioEquipo repositorioEquipo) {
         this.repositorioTorneo = repositorioTorneo;
+        this.repositorioEquipo = repositorioEquipo;
     }
 
     private void validarQueNoSeSuperponganFechas(TorneoVirtual torneoNuevo) throws FechasSuperpuestasException {
@@ -39,11 +42,12 @@ public class ServicioTorneoImpl implements ServicioTorneo {
 
 
     @Override
-    public void crearTorneo(TorneoVirtual torneo) throws FechaIncoherenteException {
+    public void crearTorneo(TorneoVirtual torneo) throws FechaIncoherenteException, FechasSuperpuestasException {
         if (torneo.getFechaFin().isBefore(torneo.getFechaInicio())) {
             throw new FechaIncoherenteException("La fecha de finalizacion del torneo debe ser posterior a la fecha de inicio");
 
         }
+        validarQueNoSeSuperponganFechas(torneo);
         repositorioTorneo.guardarTorneo(torneo);
     }
 
@@ -60,9 +64,15 @@ public class ServicioTorneoImpl implements ServicioTorneo {
 
 
     @Override
-    public void eliminarTorneo(Long id) {
-        Torneo torneo = buscarTorneoPorId(id);
+    public void eliminarTorneo(Long id) throws NoSePuedeEliminarUnTorneoSiTieneEquiposAsociadosException {
+
+        if (repositorioEquipo.existeEquipoEnTorneo(id)){
+            throw new NoSePuedeEliminarUnTorneoSiTieneEquiposAsociadosException("No se puede eliminar el torneo ya que tiene equipos asociados");
+        }
+       Torneo torneo = buscarTorneoPorId(id);
+
         repositorioTorneo.eliminarTorneo(torneo);
+
     }
 
     @Override
